kernel: move kernel object APIs to own header
Part of an effort to break up the gigantic kernel.h.
