+++ include/kernel_includes.h
@@ -43,5 +43,6 @@
 #include <irq.h>
 #include <sys/thread_stack.h>
 #include <app_memory/mem_domain.h>
+#include <sys/kobject.h>
 
 #endif /* ZEPHYR_INCLUDE_KERNEL_INCLUDES_H_ */
