+++ subsys/testsuite/ztest/include/syscalls/kobject.h
@@ -1,0 +1,5 @@
+/*
+ * Copyright (c) 2020 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
